python-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.1
onnxruntime==1.17.0
pydantic==2.6.3
numpy==1.26.3
pandas==2.2.1
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.4
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0